Refactoring UserProfile Area - prevent users from editing other user objects that are not their own - clean up code

+++ bapccanada/user/views.py
@@ -3,7 +3,7 @@ from django.contrib.auth.forms import PasswordChangeForm
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.models import User
 from django.db.models import Sum
-from django.http import JsonResponse
+from django.http import JsonResponse, Http404, HttpResponseNotFound
 from django.shortcuts import get_object_or_404
 # Create your views here.
 from django.urls import reverse_lazy
@@ -12,6 +12,7 @@ from django.views.generic import TemplateView, DeleteView
 from build.models import Build
 from products.models import Review
 from user.forms import BiographyForm, AvatarForm, ClickSettingsForm, PrivacySettingsForm, EmailSettingsForm, ReviewForm
+from user.models import UserProfile
 
 
 class BaseProfileView(TemplateView):
@@ -34,23 +35,22 @@ class ProfileView(BaseProfileView):
     title_name = 'Profile'
 
     def post(self, request, *args, **kwargs):
-        context = self.get_context_data()
-        if 'biography' in request.POST:
-            context['biography_form'] = BiographyForm(request.POST, instance=context['browse_user'].userprofile)
-            if context['biography_form'].is_valid():
-                context['biography_form'].save(commit=True)
-        if 'avatar' in request.FILES:
-            context['avatar_form'] = AvatarForm(files=request.FILES,
-                                                instance=context['browse_user'].userprofile)
-            if context['avatar_form'].is_valid():
-                context['avatar_form'].save(commit=True)
-        if 'avatar-clear' in request.POST:
-            context['avatar_form'] = AvatarForm(request.POST,
-                                                instance=context['browse_user'].userprofile)
-            if context['avatar_form'].is_valid():
-                context['avatar_form'].save(commit=True)
+        if request.user.pk == self.browse_user.pk:
+            if 'biography' in request.POST:
+                form = BiographyForm(request.POST, instance=request.user.userprofile)
+                if form.is_valid():
+                    form.save(commit=True)
+            if 'avatar' in request.FILES:
+                form = AvatarForm(files=request.FILES,
+                                  instance=self.request.user.userprofile)
+                if form.is_valid():
+                    form.save(commit=True)
+            if 'avatar-clear' in request.POST:
+                form = AvatarForm(request.POST,
+                                  instance=self.request.user.userprofile)
+                if form.is_valid():
+                    form.save(commit=True)
         return self.get(request, *args, **kwargs)
-        # return super(ProfileView, self).render_to_response(context)
 
     def get_context_data(self, **kwargs):
         context = super(ProfileView, self).get_context_data(**kwargs)
@@ -66,17 +66,17 @@ class PreferencesView(BaseProfileView):
     title_name = 'Preferences'
 
     def post(self, request, *args, **kwargs):
-        context = self.get_context_data()
-        clickSettingsForm = ClickSettingsForm(request.POST, instance=context['browse_user'].userprofile.clicksettings)
-        if clickSettingsForm.is_valid():
-            clickSettingsForm.save()
-        emailSettingsForm = EmailSettingsForm(request.POST, instance=context['browse_user'].userprofile.emailsettings)
-        if emailSettingsForm.is_valid():
-            emailSettingsForm.save()
-        privacySettingsForm = PrivacySettingsForm(request.POST,
-                                                  instance=context['browse_user'].userprofile.privacysettings)
-        if privacySettingsForm:
-            privacySettingsForm.save()
+        if request.user.pk == self.browse_user.pk:
+            clickSettingsForm = ClickSettingsForm(request.POST, instance=self.request.user.userprofile.clicksettings)
+            if clickSettingsForm.is_valid():
+                clickSettingsForm.save()
+            emailSettingsForm = EmailSettingsForm(request.POST, instance=self.request.user.userprofile.emailsettings)
+            if emailSettingsForm.is_valid():
+                emailSettingsForm.save()
+            privacySettingsForm = PrivacySettingsForm(request.POST,
+                                                      instance=self.request.user.userprofile.privacysettings)
+            if privacySettingsForm:
+                privacySettingsForm.save()
         return self.get(request, *args, **kwargs)
 
     def get_context_data(self, **kwargs):
@@ -88,42 +88,38 @@ class PreferencesView(BaseProfileView):
         return context
 
 
-class CommentsView(BaseProfileView):
+class ReviewView(BaseProfileView):
     template_name = 'comments.html'
     title_name = 'Comments'
 
     def get_context_data(self, **kwargs):
-        context = super(CommentsView, self).get_context_data(**kwargs)
+        context = super(ReviewView, self).get_context_data(**kwargs)
         context['reviews'] = context['browse_user'].userprofile.review_set.all()[:10]
         return context
 
 
-class CommentsDeleteView(DeleteView):
-    model = Review
-    template_name = 'comments_delete.html'
-
-    def get_success_url(self):
-        return JsonResponse({"test": "test"})
-
-
 def review_delete(request, **kwargs):
-    review = get_object_or_404(Review, pk=kwargs['pk'])
+    try:
+        review = request.user.userprofile.review_set.get(pk=kwargs['pk'])
+    except Review.DoesNotExist:
+        raise Http404()
     if request.method == 'POST':
-        review.delete()
-        data = {
-            "was_deleted": True,
-            "pk": kwargs['pk']
-        }
-
-        return JsonResponse(data)
-    else:
-        return JsonResponse({
-            'reponse': 'Not a valid delete request'
-        })
+        if review.user.pk == request.user.pk:
+            review.delete()
+            data = {
+                "was_deleted": True,
+                "pk": kwargs['pk']
+            }
+            return JsonResponse(data)
+    return Http404()
 
 
 def review_update(request, **kwargs):
-    review = get_object_or_404(Review, pk=kwargs['pk'])
+    try:
+        review = request.user.userprofile.review_set.get(pk=kwargs['pk'])
+    except Review.DoesNotExist:
+        raise Http404()
+
     if request.method == 'POST':
         form = ReviewForm(data={
             "content": request.POST['data']
@@ -138,9 +134,7 @@ def review_update(request, **kwargs):
                     "content": review.content
                 }
             })
-    return JsonResponse({
-        'reponse': 'Not a valid update request'
-    })
+    return Http404()
 
 
 class BuildsView(BaseProfileView):
@@ -158,7 +152,7 @@ class BuildsView(BaseProfileView):
                 context['build'] = get_object_or_404(self.browse_user.userprofile.build_set, pk=kwargs['pk'])
                 context['component_list'] = Build.get_component_dict(context['build'])
 
-            else: # default to display first build
+            else:  # default to display first build
                 context['build'] = self.browse_user.userprofile.build_set.first()
                 context['component_list'] = Build.get_component_dict(context['build'])
         return context
